% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{fuse_jaccardmat}
\alias{fuse_jaccardmat}
\title{get a fusion matrix of two jaccard matrix}
\usage{
fuse_jaccardmat(jaccard.mat1, jaccard.mat2, cores)
}
\arguments{
\item{jaccard.mat1}{a jaccard matrix}

\item{jaccard.mat2}{a jaccard matrix}

\item{cores}{cores}
}
\value{
a fusion jaccard matrix
}
\description{
get a fusion matrix of two jaccard matrix
}
