% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{fuse_ref_jm}
\alias{fuse_ref_jm}
\title{to get a whole reference combined input and built-in reference}
\usage{
fuse_ref_jm(jaccard.mat1, jaccard.mat2, info_reference, cores)
}
\arguments{
\item{jaccard.mat1}{jaccard matrix}

\item{jaccard.mat2}{jaccard matrix}

\item{info_reference}{data.frame}

\item{cores}{cores}
}
\value{
jaccard matrix
}
\description{
to get a whole reference combined input and built-in reference
}
