% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_cluster_acc}
\alias{get_cluster_acc}
\title{Calculate cell type composition of each cluster}
\usage{
get_cluster_acc(sim_mat, clusters, types)
}
\arguments{
\item{sim_mat}{A similarity matrix}

\item{clusters}{Clusters}

\item{types}{Cell types}
}
\value{
List of composition of cell types for each cluster, and proportion of the largest cell type in the cluster
}
\description{
Calculate cell type composition of each cluster
}
