% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_cluster_ratio}
\alias{get_cluster_ratio}
\title{alculate cell type composition of each cluster for all layers}
\usage{
get_cluster_ratio(sim_mat, info, layer)
}
\arguments{
\item{sim_mat}{A similarity matrix}

\item{info}{A information dataframe from Suerat object list}

\item{layer}{Layer}
}
\value{
List of composition of cell types for each cluster, and proportion of the largest cell type in the cluster for all layers
}
\description{
alculate cell type composition of each cluster for all layers
}
