% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_info}
\alias{get_info}
\title{Get a information dataframe from Suerat object list}
\usage{
get_info(seulist, type, clustername)
}
\arguments{
\item{seulist}{A list of Seurat objects}

\item{type}{The type of Seurat: 'query' or 'reference'}
}
\value{
A information dataframe of Suerat object list
}
\description{
Get a information dataframe from Suerat object list
}
