% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_jaccardmat}
\alias{get_jaccardmat}
\title{calculate jaccard matrix of Seurat object}
\usage{
get_jaccardmat(seulist, type, cores)
}
\arguments{
\item{seulist}{a list of Seurat object}

\item{type}{query or reference}

\item{cores}{cores}
}
\value{
jaccard matrix
}
\description{
calculate jaccard matrix of Seurat object
}
