% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_layers}
\alias{get_layers}
\title{Get layer information}
\usage{
get_layers(info_reference, info_layer)
}
\arguments{
\item{info_reference}{A dataframe}

\item{info_layer}{A dataframe of layer information}
}
\value{
A dataframe includes layer information
}
\description{
Get layer information
}
