% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_module}
\alias{get_module}
\title{Selecting an appropriate number of modules and defining the module identity for each cluster}
\usage{
get_module(sim_mat, acc_list_sta, m_thres = 0.8, p_thres = 0.8, min_nm = NULL)
}
\arguments{
\item{sim_mat}{A similarity matrix}

\item{acc_list_sta}{A list, where each element corresponds to cell types of a cluster}

\item{m_thres}{The minimum purity threshold that each module satisfies}

\item{p_thres}{Proportion of modules meeting the m_thres threshold}

\item{min_nm}{Minimum number of modules}
}
\value{
Number of modules and module identity for each cluster
}
\description{
Selecting an appropriate number of modules and defining the module identity for each cluster
}
