% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_sim_mat}
\alias{get_sim_mat}
\title{Calculate similarity matrix of two ordered gene list}
\usage{
get_sim_mat(gene_list1, gene_list2 = NULL, cores)
}
\arguments{
\item{cores}{Number of CPU for running}

\item{genelist1}{Gene list 1}

\item{genelist2}{Gene list 2}
}
\value{
Similarity matrix
}
\description{
Calculate similarity matrix of two ordered gene list
}
