% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{merged_weighted}
\alias{merged_weighted}
\title{Calculate weight of genes afer merging clusters into module}
\usage{
merged_weighted(gene_sets, a = 0.5, score = FALSE)
}
\arguments{
\item{gene_sets}{List of ordered genes of all clusters}

\item{a}{A hyperparameter that adjusts the weights of gene order and frequency.}

\item{score}{TRUE outputs the score of each gene after merging clusters into module. By default FALSE.}
}
\value{
Ordered gene vector
}
\description{
Calculate weight of genes afer merging clusters into module
}
