% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{module_celltype}
\alias{module_celltype}
\title{Calculate cell type composition of each module}
\usage{
module_celltype(acc_list_sta, out.id, k)
}
\arguments{
\item{acc_list_sta}{Cell type composition of each cluster}

\item{out.id}{Identity of the cluster to which each cell belongs}

\item{k}{Number of clusters}
}
\value{
List of composition of cell types for each module, and proportion of the largest cell type in the module
}
\description{
Calculate cell type composition of each module
}
