% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_obj}
\alias{process_obj}
\title{Pre-process data and add corresponding order information to Seurat object}
\usage{
process_obj(seulist, seuorder = NULL, type = "query")
}
\arguments{
\item{seulist}{A list of Seurat objects or a Seurat object}

\item{seuorder}{A vector that show the order of Suerat in the list}

\item{type}{The type of Seurat: 'query' or 'reference'}
}
\value{
A list of Suerat object that has been well processed
}
\description{
Pre-process data and add corresponding order information to Seurat object
}
\examples{
query <- process_obj(query, seuorder, 'query')
}
