% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{scPlantGM}
\alias{scPlantGM}
\title{Predict cell types based on modules}
\usage{
scPlantGM(
  query,
  reference,
  topmarkers_ref = NULL,
  topmarkers_query = NULL,
  species,
  organ,
  custom = "None",
  layer_info = NA,
  layer = 0,
  p_thres = 0.8,
  m_thres = 0.8,
  x_thres = 0.2,
  ref_clustername = "seurat_clusters",
  query_clustername = "seurat_clusters",
  topn = 100,
  cores = NA
)
}
\arguments{
\item{query}{A list of query Seurat object}

\item{reference}{A list of reference Seurat object (can be ignored if custom='None')}

\item{topmarkers_ref}{A list, top markers of each cluster obtained from Seurat pipeline for reference Seurat object}

\item{topmarkers_query}{A data.frame, top markers of each cluster obtained from Seurat pipeline for query Seurat object}

\item{species}{Species of scRNA-seq data (built-in species: Arabidopis, Maize, Rice)}

\item{organ}{Organ of species \link{Built-in organs: Arabidopis(Root,Leaf,Pollen,Inflorescence), Maize(Root,Leaf), Rice(Root,Leaf)}}

\item{custom}{Custom use built-in reference or not. Users can input 'All', 'Semi' or 'None'}

\item{layer_info}{A format-fixed data.frame includes layers information (can be ignored if custom='None')}

\item{layer}{Layer of prediction: must be an int (0,1,2...)}

\item{p_thres}{A threshold for determining what level of module purity is acceptable. The higher the value, the purer the module are}

\item{m_thres}{A threshold for determining how many modules can be accepted. The higher the value, the more modules can be accepted}

\item{x_thres}{A threshold for determining whether the cell type should be ‘unassigned’}

\item{cores}{Number of CPU for running}
}
\value{
Prediction result in data.frame
}
\description{
Predict cell types based on modules
}
\examples{
result -> scPlantGM(query, reference, species = 'Maize', organ = 'Root', layer = 0, custom = 'None')
}
