% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{sim_score}
\alias{sim_score}
\title{Calculate the similarity score between two ordered gene sets}
\usage{
sim_score(set1, set2)
}
\arguments{
\item{set1}{gene set 1}

\item{set2}{gene set 2}
}
\value{
stat
}
\description{
Calculate the similarity score between two ordered gene sets
}
