% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_cluster_acc}
\alias{get_cluster_acc}
\title{get cluster accuracy}
\usage{
get_cluster_acc(jaccard_mat, clusters, types)
}
\arguments{
\item{jaccard_mat}{a jaccard matrix}

\item{clusters}{clusters}

\item{types}{type}
}
\value{
cluster accuracy
}
\description{
get cluster accuracy
}
