% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_cluster_ratio}
\alias{get_cluster_ratio}
\title{get accuracy list}
\usage{
get_cluster_ratio(jaccard.mat, info, layer)
}
\arguments{
\item{jaccard.mat}{a jaccard matrix}

\item{info}{a data.frame}

\item{layer}{layer}
}
\value{
accuracy list
}
\description{
get accuracy list
}
