% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Calculate ordered gene list of clusters based on Seurat object}
\usage{
get_gene_list(
  seulist,
  seuorder,
  clustername,
  topn,
  cores,
  topmarkers_list = NULL
)
}
\arguments{
\item{seulist}{A list of Seurat objects or a Seurat object}

\item{seuorder}{A vector that show the order of Suerat in the list}

\item{clustername}{Cluster column name used}

\item{topn}{Number of top markers of each cluster}

\item{topmarkers}{A list, top markers of each cluster obtained from Seurat pipeline}
}
\value{
Ordered gene list
}
\description{
Calculate ordered gene list of clusters based on Seurat object
}
