% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_info}
\alias{get_info}
\title{to get a information dataframe from Suerat object list}
\usage{
get_info(seulist, type)
}
\arguments{
\item{seulist}{a list of seurat objects}

\item{type}{type}
}
\value{
a dataframe
}
\description{
to get a information dataframe from Suerat object list
}
