% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_layers}
\alias{get_layers}
\title{to get layer information}
\usage{
get_layers(info_reference, info_layer)
}
\arguments{
\item{info_reference}{a data.frame}

\item{info_layer}{a data.frame of layer information}
}
\value{
a data.frame includes layer information
}
\description{
to get layer information
}
