% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_markers}
\alias{get_markers}
\title{get markers from Seurat object}
\usage{
get_markers(object, clustername, topn, cores, avg.log2fc.threshold = 0.25)
}
\arguments{
\item{object}{a Seurat object}

\item{clustername}{Cluster column name used}

\item{topn}{Number of top markers of each cluster}

\item{cores}{Number of CPU for running}
}
\value{
markers
}
\description{
get markers from Seurat object
}
