% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_markers}
\alias{get_markers}
\title{get markers from Seurat object}
\usage{
get_markers(rds)
}
\arguments{
\item{rds}{a Seurat object}
}
\value{
markers
}
\description{
get markers from Seurat object
}
