% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{get_module}
\alias{get_module}
\title{get modules}
\usage{
get_module(jaccard_mat, acc_list_sta, p_thres = 0.8, m_thres = 0.9)
}
\arguments{
\item{jaccard_mat}{a jaccard matrix}

\item{acc_list_sta}{accuracy list}

\item{p_thres}{a threshold}

\item{m_thres}{a threshold}
}
\value{
module
}
\description{
get modules
}
