% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objectives.R
\name{module_celltype}
\alias{module_celltype}
\title{get connection between modules and cell types}
\usage{
module_celltype(acc_list_sta, out.id, k)
}
\arguments{
\item{acc_list_sta}{accuracy list}

\item{out.id}{out.id}

\item{k}{k}
}
\value{
a list of accuracy list
}
\description{
get connection between modules and cell types
}
