% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_obj}
\alias{process_obj}
\title{to pre-process Suerat objects into a fixed format}
\usage{
process_obj(seulist, seuorder, type = "query")
}
\arguments{
\item{seulist}{one/a list of Seurat objects}

\item{seuorder}{a vector that show the order of Suerat in the list. Strongly recommend using 'Sample' to present the order}

\item{type}{the type of Suerat: 'query' or 'reference'}
}
\value{
a list of Suerat that has been well processed
}
\description{
to pre-process Suerat objects into a fixed format
}
\examples{
query <- process_obj(query, seuorder, 'query')
}
