% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{scPlantGM}
\alias{scPlantGM}
\title{Predict cell types based on modules}
\usage{
scPlantGM(
  query,
  reference,
  species,
  organ,
  custom = "None",
  layer_info = NA,
  layer = 0,
  p_thres = 0.8,
  m_thres = 0.9,
  x_thres = 0.1,
  cores = NA
)
}
\arguments{
\item{query}{a list of query Seurat object}

\item{reference}{a list of reference Seurat object (can be ignored if custom='None')}

\item{species}{species of scRNA-seq data (built-in species: Arabidopis, Maize, Rice)}

\item{organ}{organ of species \link{Built-in organs: Arabidopis(Root,Leaf,Pollen,Inflorescence), Maize(Root,Leaf), Rice(Root,Leaf)}}

\item{custom}{custom use built-in reference or not. Users can input 'All', 'Semi' or 'None'}

\item{layer_info}{a format-fixed data.frame includes layers information (can be ignored if custom='None')}

\item{layer}{layer of prediction: must be an int (0,1,2...)}

\item{p_thres}{a threshold for determining what level of module purity is acceptable. The higher the value, the purer the module are}

\item{m_thres}{a threshold for determining how many modules can be accepted. The higher the value, the more modules can be accepted}

\item{x_thres}{a threshold for determining whether the cell type should be ‘unknown’}

\item{cores}{numbers of CPU for running}
}
\value{
prediction result in data.frame
}
\description{
Predict cell types based on modules
}
\examples{
result -> scPlantGM(query, reference, species = 'Maize', organ = 'Root', layer = 0, custom = 'None')
}
